fix(cmd): context not correctly shared across subcommands  (#161)
Currently, some of the DevPod CLI sub-commands do not honor global flags `--context` and `--provider`.

For instance:
```shell
$ devpod context create box
$ devpod ide use intellij --context default
```
sets the default ide on the 'box' context and not on the 'default' context, as can be verified with

```shell
$ devpod ide list
$ devpod context use default
$ devpod ide list
```
which is wrong ;)

The fix seems to be to move a '*' from the structure creation to the structure declaration, the way most of the sub-commands do it.

I have only a vague idea of what this all means, this being the first Go code I ever touched, but it does bring up C(++) memories from the last millennium...

I applied this adjustment to all the 13 commands that seem to have misplaced that crucial '*':
- context create/delete/list/set-options/use
- ide list/options/set-options/use
- provider list/list-available/set-options/use

It is not clear to me how this was not detected earlier ;)

Signed-off-by: Leonid Dubinsky <dub@podval.org>